Separate puppeteer from api module
